Crate rmqtt

Structs

MqttSession
This structure represents a mqtt session. It can be created using MqttSessionBuilder.

MqttSessionBuilder
A structure that can be used to connect to mqtt-broker and build a session.

ReceivedMessage
Message received from broker. It contains information about topic to which this message has been sent and payload of this message as bytes vector.

SubscribeResult
Result of subscribe method.

UnsubscribeResult
Result of unsubscribe method.

Enums

ConnectFailed
Errors that can occure while trying to connect to mqtt broker.

PublishResult
PublishResult of messages published with quality of service 0 is always Ready (the publishing is complete). Other messages must be acknowledged by broker. In this case the result of publishing must by checked using method await_publish_completion. This method uses packet_id contained in PublishResult::NotComplete, returned by publish method, as its argument.

ReceiveFailed
Possible results of the MqttSession.await_* methods.