Struct rmqtt::MqttSession

pub struct MqttSession {
    // some fields omitted
}

This structure represents a mqtt session. It can be created using MqttSessionBuilder.

Methods

impl MqttSession

fn reconnect(&mut self) -> bool

When connection is lost, this method can be used to reconnect.

fn subscribe(&mut self, topic: &str, qos: u8) -> SubscribeResult

This method sends a SUBSCRIBE packet to mqtt-broker.

Note

Calling this method does not mean that the topic is immediately subscribed. To make sure that the client subscribed successfully to topic method await_subscribe_completed should be used.

fn unsubscribe(&mut self, topic: &str) -> UnsubscribeResult

This method sends an UNSUBSCRIBE packet to mqtt-broker.

Note

Calling this method does not mean that the topic is immediately unsubscribed. To make sure that the client unsubscribed the topic successfully method await_unsubscribe_completed should be used.

fn publish(&mut self, topic: &str, payload: Vec<u8>, qos: u8) -> PublishResult

This method sends a PUBLISH packet to mqtt-broker.

Note

Calling this method does not mean that a message is immediately published. To make sure that the message has been successfully published to the topic, method await_publish_completion should be used.

fn disconnect(&mut self)

This method sends DISCONNECT packet to mqtt-broker.

fn await_subscribe_completed(&mut self, subscribe_packet_id: i16, timeout: Option<Duration>) -> Result<u8, ReceiveFailed>

Method waits (blocks the execution) until the topic subscription is completed.

fn await_unsubscribe_completed(&mut self, unsubscribe_packet_id: i16, timeout: Option<Duration>) -> Result<bool, ReceiveFailed>

Method waits (blocks the execution) until the topic is successfully unsubscribed.

fn await_new_message(&mut self, timeout: Option<Duration>) -> Result<ReceivedMessage, ReceiveFailed>

Method waits (blocks the execution) on a new message.

fn await_publish_completion(&mut self, publish_packet_id: i16, timeout: Option<Duration>) -> Result<PublishResult, ReceiveFailed>

Method waits (blocks the execution) until the message publishing is completed.

Note

timeout considers only second-part of the Duration