Struct rmqtt::MqttSessionBuilder

pub struct MqttSessionBuilder {
    // some fields omitted
}

A structure that can be used to connect to mqtt-broker and build a session.

Examples

use rmqtt::MqttSessionBuilder;

match MqttSessionBuilder::new("test-client", "localhost:1883")
        .credentials("user", "password")
        .keep_alive(120)
        .connect() {
    Ok(ref mut mqtt_session) => {
        // do somthing with session
    },
    Err(message) => {
        // session creation failed (cause can be found in message)
    }
}

Methods

impl MqttSessionBuilder

fn new(client_id: &str, host: &str) -> MqttSessionBuilder

Creates a new MqttSessionBuilder.

fn credentials(self, user: &str, password: &str) -> MqttSessionBuilder

This method can be used to specify the username and password that should be used to authenticated the client when connecting to the mqtt broker.

If this method is not called then the MqttSessionBuilder tries to create a connection without authntication.

fn will_message(self, will_topic: &str, will_content: &str, will_qos: u8, will_retain: bool) -> MqttSessionBuilder

This method can be used to define the Will Message. Will Message is (acording to the mqtt 3.1.1 documentation) a message that will be published to the in this method specified topic when the unexpected connection lost between client and broker occures.

If this method is not called then no Will Message is defined.

fn keep_alive(self, keep_alive: i16) -> MqttSessionBuilder

This method can be used to define how often should the broker expect a keep alive message from client. The time period is defined in seconds.

fn clean_session(self) -> MqttSessionBuilder

Call of this method indicates that a clean session should be initialized.

fn connect(self) -> Result<MqttSession, ConnectFailed>

This method can be used to initialize the connetion between broker and client using parameters defined in this MqttSessionBuilder.